+++ biofilter/db/models/model_diseases.py
@@ -1,6 +1,6 @@
 from biofilter.db.base import Base
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, Integer, String, ForeignKey, Text
+from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey, Text
 
 
 class DiseaseGroup(Base):
@@ -96,7 +96,7 @@ class DiseaseMaster(Base):
 
     # Links to the central Entity table
     entity_id = Column(
-        Integer, ForeignKey("entities.id", ondelete="CASCADE"), nullable=False
+        BigInteger, ForeignKey("entities.id", ondelete="CASCADE"), nullable=False
     )
     entity = relationship("Entity", passive_deletes=True)
